move all rule related methods from RODSConn to RC_rule trait
